add run_ray_detached_actors function
