+++ skyrl/tinker/entrypoints/main.py
@@ -1,10 +1,29 @@
+import time
 import argparse
 import ray
 
 from skyrl.tinker.config import APIConfig, EngineConfig, SkyRLTxConfig, add_model
 from skyrl.backends.jax import JaxBackendConfig
-from skyrl.tinker.ray_actors import TinkerAPIActor, TinkerEngineActor
+from skyrl.tinker.ray import run_ray_detached_actors
 from skyrl.utils.log import logger
+
+import tinker
+import numpy as np
+from tinker import types
+
+
+def process_example(example, tokenizer):
+    prompt = f"English: {example['input']}\nPig Latin:"
+    prompt_tokens = tokenizer.encode(prompt, add_special_tokens=True)
+    completion_tokens = tokenizer.encode(f" {example['output']}\n\n", add_special_tokens=False)
+
+    tokens = prompt_tokens + completion_tokens
+    weights = [0] * len(prompt_tokens) + [1] * len(completion_tokens)
+
+    return types.Datum(
+        model_input=types.ModelInput.from_ints(tokens=tokens[:-1]),
+        loss_fn_inputs=dict(weights=weights[1:], target_tokens=tokens[1:])
+    )
 
 def main():
     logger.info("Starting entrypoint...")
@@ -26,31 +45,31 @@ def main():
     
     logger.info(f"Initializing Ray with address: {config.engine.ray_address or 'local'}")
     ray.init(address=config.engine.ray_address)
+    run_ray_detached_actors(config)
 
-    logger.info("Creating STRICT_PACK placement group to colocate API and Engine actors...")
-    from ray.util.placement_group import placement_group
-    pg = placement_group([{"CPU": 1}, {"CPU": 1}], strategy="STRICT_PACK")
-    ray.get(pg.ready())
+    time.sleep(100)
 
-    logger.info(f"Starting Tinker API Actor on {config.api.host}:{config.api.port} (Detached)...")
-    api_actor = TinkerAPIActor.options(
-        placement_group=pg,
-        placement_group_bundle_index=0,
-        name="tinker_api",
-        lifetime="detached"
-    ).remote(config.engine)
-    api_actor.run.remote(config.api.host, config.api.port)
+    service_client = tinker.ServiceClient(base_url="http://localhost:8000", api_key="tml-dummy")
+    training_client = service_client.create_lora_training_client(base_model="Qwen/Qwen3-0.6B")
+    tokenizer = training_client.get_tokenizer()
 
-    logger.info("Starting Tinker Engine Actor (Detached)...")
-    engine_actor = TinkerEngineActor.options(
-        placement_group=pg,
-        placement_group_bundle_index=1,
-        name="tinker_engine",
-        lifetime="detached"
-    ).remote(config.engine)
-    engine_actor.run.remote()
+    # Training examples
+    examples = [
+        {"input": "banana split", "output": "anana-bay plit-say"},
+        {"input": "quantum physics", "output": "uantum-qay ysics-phay"},
+        {"input": "coding wizard", "output": "oding-cay izard-way"},
+    ]
 
-    logger.info("Ray actors started in detached mode. They will keep running. You can now run your training script.")
+    processed = [process_example(ex, tokenizer) for ex in examples]
+
+    # Training loop
+    for _ in range(6):
+        fwdbwd = training_client.forward_backward(processed, "cross_entropy").result()
+        training_client.optim_step(types.AdamParams(learning_rate=1e-4)).result()
+
+        logprobs = np.concatenate([o['logprobs'].tolist() for o in fwdbwd.loss_fn_outputs])
+        weights = np.concatenate([e.loss_fn_inputs['weights'].tolist() for e in processed])
+        print(f"Loss: {-np.dot(logprobs, weights) / weights.sum():.4f}")
 
 
 if __name__ == "__main__":
